pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "wpsrt"
version = "0.1.0"
description = ""
authors = ["jpk <jpk@goatpr0n.de>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
isort = "^5.10.1"

[tool.poetry.scripts]
wpsrt = "wpsrt.cli:sort"
